+++ src/CacheBuster.ts
@@ -85,7 +85,7 @@ export class CacheBuster extends EventEmitter {
 
   protected referenceRx =
     /((")(([^"\s]+)\.([a-zA-Z0-9]+))")|((')(([^'\s]+)\.([a-zA-Z0-9]+))')|# sourceMappingURL=(([^\s]+)\.([a-zA-Z0-9]+))/g;
-  protected externalReferenceRx = /^([a-zA-Z0-9]+:|\/\/)/;
+  protected externalReferenceRx = /^([a-zA-Z][a-zA-Z0-9+.\-]*:|\/\/)/;
   protected baseRx = /<base[^<>]+href\s*=\s*"([^"]+)"[^<>]*>/i;
   protected extensionRx = /(?<=\.)[^.]+$/;
 
@@ -278,7 +278,13 @@ export class CacheBuster extends EventEmitter {
   }
 
   isResolvableReference(reference: string) {
-    return !reference.match(this.externalReferenceRx);
+    return (
+      !this.isExternalReference(reference) && !this.excludeGlob.match(reference)
+    );
+  }
+
+  isExternalReference(reference: string) {
+    return !!reference.match(this.externalReferenceRx);
   }
 
   resolveReference(base: string, reference: string) {