+++ src/bin.ts
@@ -3,4 +3,4 @@ import { CacheBusterCli } from './CacheBusterCli';
 
 const cli = new CacheBusterCli();
 
-cli.run(process.argv.slice(2));
+cli.run(process.argv.slice(2)).then((exitCode) => process.exit(exitCode));